Add default tile size to app preferences
All newly created tabs will have the defual tile size specified in the app preferences
