+++ app/src/main/java/edu/wpi/first/shuffleboard/app/prefs/AppPreferences.java
@@ -6,7 +6,9 @@ import com.google.common.collect.ImmutableList;
 import edu.wpi.first.shuffleboard.app.theme.DefaultThemes;
 import edu.wpi.first.shuffleboard.app.theme.Theme;
 
+import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.Property;
+import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleObjectProperty;
 
 /**
@@ -17,6 +19,7 @@ public final class AppPreferences {
 
   private final Property<Theme> theme
       = new SimpleObjectProperty<>(this, "Theme", DefaultThemes.LIGHT);
+  private final DoubleProperty defaultTileSize = new SimpleDoubleProperty(this, "defaultTileSize", 128);
 
   @VisibleForTesting
   static AppPreferences instance = new AppPreferences();
@@ -30,7 +33,8 @@ public final class AppPreferences {
    */
   public ImmutableList<Property<?>> getProperties() {
     return ImmutableList.of(
-        theme
+        theme,
+        defaultTileSize
     );
   }
 
@@ -46,4 +50,15 @@ public final class AppPreferences {
     this.theme.setValue(theme);
   }
 
+  public double getDefaultTileSize() {
+    return defaultTileSize.get();
+  }
+
+  public DoubleProperty defaultTileSizeProperty() {
+    return defaultTileSize;
+  }
+
+  public void setDefaultTileSize(double defaultTileSize) {
+    this.defaultTileSize.set(defaultTileSize);
+  }
 }
